The Choose Category Screen
This class shows the user a frame with various buttons that say what
category each chooses. Only the categories Game of Thrones and
Astronomy play sound as you play. The Game of Thrones category randomly
plays one of two songs (Rains of Castamere or the theme song) and
Astronomy plays the Interstellar theme.